# openapi.yaml
openapi: 3.1.0
info:
  title: Scrutineer Skill API
  version: 0.4.0
  description: |
    A narrow HTTP surface that claude-code skills use while they run. The
    goal is enough to let a skill orchestrate other skills and read prior
    scan outputs without pushing that logic into scrutineer's Go code.

    Every scrutineer skill scan is issued a bearer token that is staged into
    the workspace context.json. Skills authenticate with that token. A token
    is only valid while its originating scan is running, and only grants
    access to the repository the scan was issued against.

    The `/v1/*` paths are an unauthenticated, localhost-only export surface
    intended for external scripts (jq, cron, CI dumps). They emit JSONL and
    rely on the same host-header check as the browser UI; see threatmodel.md.
servers:
  - url: http://127.0.0.1:8080/api
    description: Local scrutineer server
security:
  - bearerAuth: []
paths:
...
  /repositories/{id}/scans:
    get:
      summary: List scans for a repository
      operationId: listRepositoryScans
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, paused, done, failed, skipped, cancelled]
        - name: skill
          in: query
          schema: { type: string }
          description: Filter by skill name (matches Scan.skill_name).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Scan' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /scans/{id}:
    get:
      summary: Fetch a scan with its report (if complete)
      operationId: getScan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScanDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/scans:
    get:
      summary: Export scans across all repositories as JSONL
      description: |
        Returns one JSON object per line (`application/x-ndjson`). No bearer
        token; localhost-only. Each line is a `ScanExport` carrying the full
        set of `db.Scan` columns minus `api_token` (the per-scan bearer must
        not leak through an unauthenticated channel).
      operationId: exportScans
      security: []
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [jsonl] }
        - name: status
          in: query
          schema: { type: string, enum: [queued, running, paused, done, failed, skipped, cancelled] }
        - name: skill
          in: query
          schema: { type: string }
      responses:
        '200':
          description: JSONL stream of ScanExport objects, one per line.
          content:
            application/x-ndjson:
              schema: { $ref: '#/components/schemas/ScanExport' }
        '400':
          description: Unsupported format query value, or `encrypt`/`scope`/`include` requested (only valid on per-repository bundle exports).
        '403':
          description: Host header is not localhost
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RepositoryID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Token does not grant access to this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Scan:
      type: object
      required: [id, kind, status, repository_id]
      properties:
        id:            { type: integer, format: int64 }
        repository_id: { type: integer, format: int64 }
        kind:          { type: string }
        status:        { type: string, enum: [queued, running, paused, done, failed, skipped, cancelled] }
        model:         { type: string }
        commit:        { type: string }
        rescan_mode:
          type: string
          enum: [full, diff]
          description: Requested or resolved scan mode.
        diff_base_scan_id:
          type: integer
          format: int64
          nullable: true
          description: Baseline scan used for a diff rescan.
        diff_base_commit:
          type: string
          description: Baseline commit used to generate the staged diff.
        diff_threat_model_scan_id:
          type: integer
          format: int64
          nullable: true
          description: Prior threat-model scan staged as `old_threat_model.json`, when available.
        diff_stats:
          type: string
          description: JSON metadata for staged diff inputs, including changed files and size limits.
        coverage:
          type: string
          description: JSON coverage metadata, including diff fallback reasons and threat-model update decisions.
        skill_name:           { type: string }
        skill_version:        { type: integer, description: Snapshot of the Skill row version at enqueue time. }
        skill_schema_version: { type: integer, description: Snapshot of metadata.scrutineer.version at enqueue time. }
        started_at:    { type: string, format: date-time, nullable: true }
        finished_at:   { type: string, format: date-time, nullable: true }
        max_turns_hit: { type: boolean, description: True when the scan completed with partial output after hitting the Claude max-turns cap; retry resumes the saved session when available. }
        refusal_audit_warning: { type: boolean, description: True when a security-deep-dive refusal audit reports declined or skipped analysis. }
        error:         { type: string }
    ScanDetail:
      allOf:
        - $ref: '#/components/schemas/Scan'
        - type: object
          properties:
            report: { type: string, description: Raw output report (may be JSON text). }
            refusal_audit: { type: string, description: Structured post-report security-deep-dive refusal audit. }
            log:    { type: string }
...
    ScanExport:
      description: |
        Full db.Scan row used by the unauthenticated JSONL export, minus
        api_token.
      allOf:
        - $ref: '#/components/schemas/ScanDetail'
        - type: object
          properties:
            skill_id:           { type: integer, format: int64, nullable: true }
            finding_id:         { type: integer, format: int64, nullable: true }
            sub_path:           { type: string }
            cost_usd:           { type: number }
            turns:              { type: integer }
            input_tokens:       { type: integer }
            output_tokens:      { type: integer }
            cache_read_tokens:  { type: integer }
            cache_write_tokens: { type: integer }
            prompt:             { type: string }
            findings_count:     { type: integer }
            created_at:         { type: string, format: date-time }
            updated_at:         { type: string, format: date-time }
...